requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4

# Django and extensions
Django==4.2.5
djangorestframework==3.14.0
django-crispy-forms==2.0
django-allauth==0.54.0

# FastAPI and extensions
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.11.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
pytest==7.4.0
black==23.7.0
isort==5.12.0

# Frontend
Pillow==10.0.0
